# tianming-web/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentSessionUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "isArchived": {
            "nullable": true,
            "type": "boolean"
          },
          "title": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NovelAgentWeb",
    "version": "1.0"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/agent/session/{sessionId}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Agent"
        ]
      },
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSessionUpdateRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSessionUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSessionUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Agent"
        ]
      }
    },
...
    "/api/agent/sessions/{sessionId}/resume": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Agent"
        ]
      }
    },
    "/api/agent/sse/{sessionId}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "afterEventId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Agent"
        ]
      }
    },
...
  }
}